==> Proyecto 1/notebook/02. Análisis_Exploratorio.ipynb <==
# Análisis Exploratorio

import pandas as pd
import sidetable

A continuación vamos a importar el CSV que ya habíamos creado mergeando el CSV inicial de Kagle, y el CSV creado utilizando Selenium

mergeado_inner = pd.read_csv ("../Data/01.3 Datos_enriquecidos_.csv")
mergeado_inner.head(1)

# Exploración inicial de nuestros datos

### Tipado de Datos

mergeado_inner.columns

!!!! OJO !!!! Vemos como los nombres no son homogéneos y eso nos puede dificultar nuestro trabajo. Habrá que ponerkos en minúscula, cambiar espacios por "_", y eliminar las columnas que no nos aporten valor

### Tamaño

mergeado_inner.shape

### Tipos de datos

¿Y el tipo de datos de cada una de las columnas?

mergeado_inner.dtypes

!!!! OJO !!!! No tiene mucho sentido que el número de 'baños' sea un "float", ni que los números de "Índice_poder_adqui', 'Índice_seguridad', 'Índice_sanidad', 'Índice_clima', 'Índice_costo_vida', 'Relación_precio_ingresos_propiedades', 'Índice_costo_desplazamiento', 'Índice_contaminación', 'Índice_calidad_vida' sean un "string"... Tendremos que hacer algo con ellas más adelante.



### Info

mergeado_inner.info()

No tenemos valores nulos en nuestrio DataFrame

### Principales estadísticos

mergeado_inner.describe().T

mergeado_inner.describe(include = "object").T

### Duplicados

¿Tenemos duplicados en nuestro dataframe? -> Parece que no tenemos....

mergeado_inner.duplicated().sum()

Pero, qué está pasando? Vamos a eliminar dos columnas con valores únicos "Unnamed: 0_x", "Unnamed: 0_y", y vamos a comprobar si realmente seguimos sin tener duplicados

eliminar_columns = ['Unnamed: 0.1', 'Unnamed: 0']

mergeado_inner_1 = mergeado_inner.drop(eliminar_columns, axis = 1)
mergeado_inner_1.head(2)

mergeado_inner_1.duplicated().sum()

!!!! OJO!!!!  vemos como tenemos dos valores que están duplicados en nuestro dataframe

### Sidetables

mergeado_inner_1.stb.freq(["City"])

mergeado_inner_1.stb.freq(["State"])

mergeado_inner_1[mergeado_inner_1.duplicated(keep=False)]

### Exportación de CSV mergeado con mínima limpieza inicial

Exportamos el DataFrame mergeado, y con una limpieza minimísima, a un CSV:

mergeado_inner_1.to_csv("02.1 Datos_mergeados_limpieza_inical_.csv")
