==> Proyecto 1/notebook/03. Método_de_Transformación.ipynb <==
# Método de Transformación

import pandas as pd
import sidetable
import pymongo

import folium
import seaborn as sns
import matplotlib.pyplot as plt

import sys
sys.path.append("../")
import src.operaciones as cc

### Importamos CSV

Nos descarmos el CSV que hemos creado, enriqueciendo el CSV inicial con Selenium:

df3 = pd.read_csv ("../Data/02.1 Datos_mergeados_limpieza_inical_.csv")
df3.head(2)

### Manipulación de columnas

Vamos a reescribir la columna "AD_Situation" para que únicamente muestre dos opciones: New, old. 

Vamos a reescribir la columna "State" para que únicamente muestra la primera parte de los string, es decir: WA, OR o CA.

Para ello, hemos creado una función "quitar_horas" donde haremos un split por espacio y nos quedaremos con la primera parte del string!

df3["AD_Situation"] = df3["AD_Situation"].apply(cc.quitar_horas)
df3["State"] = df3["State"].apply(cc.quitar_horas)
df3.head(1)

De nuevo vamos a reescriir la columna "State", pero para que este vez muestre los 3 Estados en su nombre completo: California, Oregón, Washington: 

Para ello hemos creado la función "cambiar"!

df3["State"] = df3.apply(lambda datos: cc.cambiar(datos["State"]),axis = 1)

Vamos a crear una nueva columna "Measuring unit" y vamos a indicar como unidad de medida "m2".

Vamos a reescribir la columna "Bulding space" para que la información se muestra en m2 en vez de sqft. Para ello lo vamos a dividir por 10.764.

Vamos a crear una columna nueva "Price/m2", donde vamos dividir la columna "Price" por la columna "Building space". 

df3['Measuring unit']= df3['Measruing unit'].replace("sqft","m2")
df3['Building space']= round(df3['Building space']/10.764, 0)
df3['Price/m2']= round((df3['Price'])/(df3['Building space']), 0)
df3.head(1)

Vamos a crear tres columnas nuevas, "Direction", "Ciudad1", "Postal_code" a partir de la columna "Address", donde vamos a hacer un split por ",", y nos vamos a quedar con las 3 primeras palabras del string:

df3[["Direction", "Ciudad1", "Postal_code"]]  = df3["Address"].str.split(",",  expand=True).iloc[:, :3]

Vamos a reescribir la columna "Direction" para que nos de la dirección y la ciudad en la misma columna. Para ello, vamos a sumar los strings de las columnas "Direction" y "Ciudad1":

df3["Direction"] = df3["Direction"] + "," + df3["Ciudad1"]
df3.head(1)

Vamos a reescribir la columna "Postal_code" para que nos de únicamente el código postal: 

Para ello, utilizando regex, vamos a quedarnos únicamente con los valores numéricos del string

df3["Postal_code"] = df3.Postal_code.str.extract("(\d+)", expand=True)
df3.head(1)

Vamos a eliminar las columna que no nos interesan por estar la información "duplicada"

eliminar_columns = ['Address', 'Unnamed: 0', 'Ciudad1', 'Measruing unit']


df3_limpio = df3.drop(eliminar_columns, axis = 1)
df3_limpio.head(1)

### Estandarización del tipado de las columnas

Vamos a ordenar las columnas de nuestro DataFrame para que tenga algo más de sentido:

df3_limpio = df3_limpio.reindex(columns=['Price', 'Currency', 'AD_Situation', 'Rooms', 'Bath', 'Building space', 'Measuring unit', 'Price/m2', 'Brokered_by', 'Direction', 'City', 'State', 'Postal_code',
       'Índice_poder_adqui', 'Índice_seguridad', 'Índice_sanidad',
       'Índice_clima', 'Índice_costo_vida',
       'Relación_precio_ingresos_propiedades', 'Índice_costo_desplazamiento',
       'Índice_contaminación', 'Índice_calidad_vida'])
df3_limpio.head(1)

A continuación, vamos a homogeneizar las columnas: pasar los nombres de las columnas a minusculas y reemplazar los espacios por "_":

diccionario = {}
for col in df3_limpio.columns:
    diccionario[col] = col.lower().strip().replace(" ", "_")

nuevas_columnas = {col : col.lower().strip().replace(" ", "_") for col in df3_limpio.columns}
nuevas_columnas

df3_limpio.rename(columns = nuevas_columnas, inplace =True)
df3_limpio.head(1)

### Tipo de Datos

Problema: La columna "baños" está en formato "float". Debemos convertirlas a integer. Las columans "índice_poder_adqui', 'índice_seguridad', 'índice_sanidad', 'índice_clima', 'índice_costo_vida', 'relación_precio_ingresos_propiedades', 'índice_costo_desplazamiento', 'índice_contaminación', 'índice_calidad_vida' están en formato "string"... Tendremos que hacer algo con ellas más adelante. Debemos convertirlas a integer:

df3_limpio.dtypes

df3_limpio["bath"] = df3_limpio["bath"].astype(int)
df3_limpio.head(2)

z = ['índice_poder_adqui', 'índice_seguridad',
       'índice_sanidad', 'índice_clima', 'índice_costo_vida',
       'relación_precio_ingresos_propiedades', 'índice_costo_desplazamiento',
       'índice_contaminación', 'índice_calidad_vida']

df3_limpio = df3_limpio.apply(lambda datos: cc.puntos2(datos, z), axis = 1)

df3_limpio.head(1)

df3_limpio.dtypes

### Duplicados

Problema: Duplicado. En el análisis exploratorio identificamos 2 duplicados:

df3_limpio.duplicated().sum()

df3_limpio[df3_limpio.duplicated() == True]

Una vez identificamos las filas duplicadas, procedemos a su eliminación:

df3_limpio.drop_duplicates(inplace = True)

df3_limpio.duplicated().sum()

### Valores Nulos

df3_limpio.isnull().sum()

df3_limpio["postal_code"] = df3_limpio["postal_code"].fillna(0)

df3_limpio.isnull().sum()

df3_limpio["postal_code"].dtypes

df3_limpio["postal_code"] = df3_limpio["postal_code"].astype(int)

df3_limpio["postal_code"].dtypes

### Principales estadísticos

Vamos a sacar los principales estadísticos a modo general, y también por estado, para ver qué insights podemos sacar

df3_limpio.describe().T

df3_limpio.describe(include = "object").T

Nos sacamos los estadísticos por Estado para ver posibles insights

df3_limpio[df3_limpio["state"] == "California"].describe().T.round()

df3_limpio[df3_limpio["state"] == "Oregón"].describe().T.round()

df3_limpio[df3_limpio["state"] == "Washington"].describe().T.round()

df3_limpio[df3_limpio["state"] == "California"].describe().T["mean"].round()

df3_limpio.columns

df_pie = pd.DataFrame(df3_limpio.groupby(['city', 'state'])
                      ["price", "building_space", "price/m2", "índice_poder_adqui"].mean()).round().reset_index()

df_pie.head(12)

### Hipótesis

Hipótesis 1: El precio de la casa está directamente relacionado con los metros cuadrados de la misma.

Hipótesis 2: Dentro de un Estado, las ciudades con un índice de Contaminación más elevado son las que tienen un índice de Salud más bajo.

Hipótesis 3: Las casas tienen el mismo número de cuartos que de cuartos de baño.

Hipótesis 4: Existe una relación directa entre la Seguridad y la Calidad de vida de una ciudad. A más seguridad, más calidad de vida.

Hipótesis 5: El Estado donde el precio medio por m2 sea mayor, será el Estado que tenga el mayor indice de poder Adquisitivo.

### Geopy

from geopy.geocoders import Nominatim

locator = Nominatim(user_agent= "myGeocoder")

df3_limpio["latidud_"] = df3_limpio.apply(lambda datos: cc.latitud_v2(
                                    datos["postal_code"]), axis = 1)

df3_limpio["longitud_"] = df3_limpio.apply(lambda datos: cc.longitud_v2(
                                    datos["postal_code"]), axis = 1)

df3_limpio.head(2)

### Exportación de CSV limpio

Exportamos el DataFrame limpio a un CSV:

df3_limpio.to_csv("03.1 Datos_limpitos_.csv")
